refactor(core): migrate rspack plugin into core
