+++ e2e/package.json
@@ -15,7 +15,6 @@
     "@rsbuild/plugin-less": "catalog:",
     "@rsdoctor/cli": "workspace:*",
     "@rsdoctor/core": "workspace:*",
-    "@rsdoctor/rspack-plugin": "workspace:*",
     "@rsdoctor/shared": "workspace:*",
     "@rspack/core": "catalog:",
     "@types/node": "catalog:",
