+++ examples/rspack-banner-minimal/package.json
@@ -18,7 +18,7 @@
     "react-dom": "catalog:"
   },
   "devDependencies": {
-    "@rsdoctor/rspack-plugin": "workspace:*",
+    "@rsdoctor/core": "workspace:*",
     "@rspack/cli": "catalog:",
     "@rspack/core": "catalog:",
     "@rspack/plugin-react-refresh": "catalog:",
