+++ examples/rspack-layers-minimal/package.json
@@ -17,7 +17,7 @@
     "react-dom": "catalog:"
   },
   "devDependencies": {
-    "@rsdoctor/rspack-plugin": "workspace:*",
+    "@rsdoctor/core": "workspace:*",
     "@rspack/cli": "catalog:",
     "@rspack/core": "catalog:",
     "@rspack/plugin-react-refresh": "catalog:",
