+++ examples/rspress-minimal/package.json
@@ -12,7 +12,7 @@
     "@rspress/core": "2.0.15"
   },
   "devDependencies": {
-    "@rsdoctor/rspack-plugin": "workspace:*",
+    "@rsdoctor/core": "workspace:*",
     "@types/node": "^24.12.3"
   }
 }
