+++ packages/core/src/rspack-plugin/index.ts
