--- packages/rspack-plugin/README.md
@@ -1,15 +1,0 @@
-# @rsdoctor/rspack-plugin
-
-An Rspack plugin for integrating Rsdoctor.
-
-## Documentation
-
-https://rsdoctor.rs/
-
-## Contributing
-
-Please read the [Contributing Guide](https://github.com/web-infra-dev/rsdoctor/blob/main/CONTRIBUTING.md).
-
-## License
-
-Rsdoctor is [MIT licensed](https://github.com/web-infra-dev/rsdoctor/blob/main/LICENSE).
