--- packages/rspack-plugin/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@rsdoctor/rspack-plugin",
-  "version": "1.5.11",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/web-infra-dev/rsdoctor",
-    "directory": "packages/rspack-plugin"
-  },
-  "license": "MIT",
-  "main": "dist/index.cjs",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      },
-      "require": {
-        "types": "./dist/index.d.cts",
-        "default": "./dist/index.cjs"
-      }
-    }
-  },
-  "type": "module",
-  "scripts": {
-    "dev": "npm run start",
-    "start": "rslib build -w",
-    "build": "rslib build",
-    "test": "rstest run"
-  },
-  "dependencies": {
-    "@rsdoctor/core": "workspace:*",
-    "@rsdoctor/shared": "workspace:*"
-  },
-  "devDependencies": {
-    "@rspack/core": "catalog:",
-    "@types/node": "catalog:",
-    "@types/tapable": "catalog:",
-    "tslib": "catalog:",
-    "typescript": "catalog:"
-  },
-  "peerDependencies": {
-    "@rspack/core": "*"
-  },
-  "peerDependenciesMeta": {
-    "@rspack/core": {
-      "optional": true
-    }
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org/"
-  }
-}
