--- packages/rspack-plugin/rslib.config.ts
@@ -1,12 +1,0 @@
-import { dualPackage } from '../../scripts/rslib.base.config';
-
-const rslibConfig = {
-  ...dualPackage,
-  ...{
-    source: {
-      tsconfigPath: 'tsconfig.build.json',
-    },
-  },
-};
-
-export default rslibConfig;
