--- packages/rspack-plugin/tsconfig.build.json
@@ -1,5 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": ["src"],
-  "exclude": ["**/node_modules"]
-}
