--- packages/rspack-plugin/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "@rsdoctor/tsconfig/base",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "src"
-  },
-  "include": ["src"],
-  "exclude": ["**/node_modules"],
-  "references": [
-    {
-      "path": "../core/tsconfig.json"
-    },
-    {
-      "path": "../shared/tsconfig.json"
-    }
-  ]
-}
